Cleanup C4267 warnings (#74)
* Fixes to avoid C4267 warnings in Widestring.cpp/h

* More C4275 Cleanup

* Updated Chunkio

* More warning C4267 fixes

* More  C4267 warnings work

* (chore) Cleaned up more warnings including remaining C4267.

* Fixes to avoid C4267 warnings in Widestring.cpp/h

* More C4275 Cleanup

* Updated Chunkio

* More warning C4267 fixes

* More  C4267 warnings work

* (chore) Cleaned up more warnings including remaining C4267.

* rebase on main

* Fixup

* Changes requested from PR

* Removed the redundant StringClass(unsigned initial_len, bool) overload in Code/wwlib/wwstring.h

* Fix indentations

* More formating cleanup

* More indent cleanup

* Attempt to force github repo to switch these files back to CRLF

* Remove comment about CRLF

* Remove comment about CRLF

* Copied files from main and re did changes to see if that fixes whitespace issues in github.

+++ Code/WWOnline/WOLNetUtilObserver.cpp
@@ -493,9 +493,9 @@ void NetUtilObserver::ProcessLadderListResults(WOL::Ladder* list, long timeStamp
 
 void NetUtilObserver::NotifyClanLadderUpdate(const UserList& users, const RefPtr<SquadData>& squad)
 	{
-	const unsigned int userCount = users.size();
+	const size_t userCount = users.size();
 
-	for (unsigned int index = 0; index < userCount; ++index)
+	for (size_t index = 0; index < userCount; ++index)
 		{
 		const RefPtr<UserData>& user = users[index];
